Moved metr.acc. to its own module
