+++ src/ParallelTemperingMonteCarlo.jl
@@ -8,9 +8,13 @@ include("RuNNer.jl")
 
 include("EnergyEvaluation.jl")
 include("MCStates.jl")
+
+
 include("InputParams.jl")
+
 #include("Initialization")
 include("MCMoves.jl")
+include("Exchange.jl")
 include("MCRun.jl")
 include("multihist.jl")
 #include("ReadSave.jl")
@@ -26,6 +30,7 @@ include("parallelrun.jl")
  @reexport using .InputParams
  #@reexport using .Initialization
  @reexport using .MCMoves
+ @reexport using Exchange
  @reexport using .MCRun
  @reexport using .Multihistogram
